+++ miaosha-common/miaosha-common-core/src/main/java/tech/tengshe789/miaosha/common/core/util/ClassUtils.java
@@ -1,5 +1,6 @@
 package tech.tengshe789.miaosha.common.core.util;
 
+import lombok.experimental.UtilityClass;
 import org.springframework.core.BridgeMethodResolver;
 import org.springframework.core.DefaultParameterNameDiscoverer;
 import org.springframework.core.MethodParameter;
@@ -13,14 +14,13 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
 
 /**
- * @program: -miaosha
- * @description: 类工具类
- * @author: tEngSHe789
- * @create: 2019-03-14 20:44
- **/
-public class ClassUtil extends org.springframework.util.ClassUtils {
-
-	private static final ParameterNameDiscoverer PARAMETER_NAME_DISCOVERER = new DefaultParameterNameDiscoverer();
+ * 类工具类
+ *
+ * @author L.cm
+ */
+@UtilityClass
+public class ClassUtils extends org.springframework.util.ClassUtils {
+	private final ParameterNameDiscoverer PARAMETERNAMEDISCOVERER = new DefaultParameterNameDiscoverer();
 
 	/**
 	 * 获取方法参数信息
@@ -29,9 +29,9 @@ public class ClassUtil extends org.springframework.util.ClassUtils {
 	 * @param parameterIndex 参数序号
 	 * @return {MethodParameter}
 	 */
-	public static MethodParameter getMethodParameter(Constructor<?> constructor, int parameterIndex) {
+	public MethodParameter getMethodParameter(Constructor<?> constructor, int parameterIndex) {
 		MethodParameter methodParameter = new SynthesizingMethodParameter(constructor, parameterIndex);
-		methodParameter.initParameterNameDiscovery(PARAMETER_NAME_DISCOVERER);
+		methodParameter.initParameterNameDiscovery(PARAMETERNAMEDISCOVERER);
 		return methodParameter;
 	}
 
@@ -42,9 +42,9 @@ public class ClassUtil extends org.springframework.util.ClassUtils {
 	 * @param parameterIndex 参数序号
 	 * @return {MethodParameter}
 	 */
-	public static MethodParameter getMethodParameter(Method method, int parameterIndex) {
+	public MethodParameter getMethodParameter(Method method, int parameterIndex) {
 		MethodParameter methodParameter = new SynthesizingMethodParameter(method, parameterIndex);
-		methodParameter.initParameterNameDiscovery(PARAMETER_NAME_DISCOVERER);
+		methodParameter.initParameterNameDiscovery(PARAMETERNAMEDISCOVERER);
 		return methodParameter;
 	}
 
@@ -56,11 +56,11 @@ public class ClassUtil extends org.springframework.util.ClassUtils {
 	 * @param <A>            泛型标记
 	 * @return {Annotation}
 	 */
-	public static <A extends Annotation> A getAnnotation(Method method, Class<A> annotationType) {
+	public <A extends Annotation> A getAnnotation(Method method, Class<A> annotationType) {
 		Class<?> targetClass = method.getDeclaringClass();
 		// The method may be on an interface, but we need attributes from the target class.
 		// If the target class is null, the method will be unchanged.
-		Method specificMethod = ClassUtil.getMostSpecificMethod(method, targetClass);
+		Method specificMethod = ClassUtils.getMostSpecificMethod(method, targetClass);
 		// If we are dealing with method with generic parameters, find the original method.
 		specificMethod = BridgeMethodResolver.findBridgedMethod(specificMethod);
 		// 先找方法，再找方法上的类
@@ -81,7 +81,7 @@ public class ClassUtil extends org.springframework.util.ClassUtils {
 	 * @param <A>            泛型标记
 	 * @return {Annotation}
 	 */
-	public static <A extends Annotation> A getAnnotation(HandlerMethod handlerMethod, Class<A> annotationType) {
+	public <A extends Annotation> A getAnnotation(HandlerMethod handlerMethod, Class<A> annotationType) {
 		// 先找方法，再找方法上的类
 		A annotation = handlerMethod.getMethodAnnotation(annotationType);
 		if (null != annotation) {