+++ miaosha-common/miaosha-common-security/src/main/java/tech/tengshe789/miaosha/common/security/aspect/InnerAspect.java
@@ -1,0 +1,41 @@
+package tech.tengshe789.miaosha.common.security.aspect;
+
+import cn.hutool.core.util.StrUtil;
+import lombok.AllArgsConstructor;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.Around;
+import org.aspectj.lang.annotation.Aspect;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.stereotype.Component;
+import tech.tengshe789.miaosha.common.core.constants.SecurityConstants;
+import tech.tengshe789.miaosha.common.security.annotation.Inner;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * @program: -miaosha
+ * @description: 服务调用不鉴权
+ * @author: tEngSHe789
+ * @create: 2019-03-26 10:51
+ **/
+@Slf4j
+@Aspect
+@Component
+@AllArgsConstructor
+public class InnerAspect {
+	private final HttpServletRequest request;
+
+	@SneakyThrows
+	@Around("@annotation(inner)")
+	public Object around(ProceedingJoinPoint point, Inner inner) {
+		String header = request.getHeader(SecurityConstants.FROM);
+		if (inner.value() && !StrUtil.equals(SecurityConstants.FROM_IN, header)) {
+			log.warn("访问接口 {} 没有权限", point.getSignature().getName());
+			throw new AccessDeniedException("Access is denied");
+		}
+		return point.proceed();
+	}
+
+}